ws_handler_fn/requirements.txt
PyJWT[crypto]
